Device Api: extract to module
We'll be adding at least one more api function here and `index.ts`
is already quite long.
